removed duplicated authentication service and authentication guard
