--- honestcash-v2/src/app/core/guards/auth-guard.service.ts
@@ -1,20 +1,0 @@
-import { Injectable } from '@angular/core';
-import { Router, CanActivate } from '@angular/router';
-
-import { AuthService } from '../services/auth.service';
-
-
-@Injectable()
-export class AuthGuardService implements CanActivate {
-  constructor(
-    public auth: AuthService,
-    public router: Router
-  ) {}
-  canActivate(): boolean {
-    if (!this.auth.getToken()) {
-      this.router.navigateByUrl('/login');
-      return false;
-    }
-    return true;
-  }
-}
