+++ honestcash-v2/src/app/core/http/token.interceptor.ts
@@ -3,16 +3,16 @@ import {
   HttpEvent, HttpInterceptor, HttpHandler, HttpRequest
 } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { AuthService } from '../services/auth.service';
+import { AuthenticationService } from '../services/authentication.service';
 
 
 @Injectable()
 export class TokenInterceptor implements HttpInterceptor {
-  private authService: AuthService;
+  private authenticationService: AuthenticationService;
   constructor(private injector: Injector) {}
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    this.authService = this.injector.get(AuthService);
-    const token: string = this.authService.getToken();
+    this.authenticationService = this.injector.get(AuthenticationService);
+    const token: string = this.authenticationService.token;
     if (token) {
       request = request.clone({
         setHeaders: {
