+++ honestcash-v2/src/app/core/services/user.service.ts
@@ -3,7 +3,7 @@ import { User } from '../../shared/interfaces';
 import { BehaviorSubject } from 'rxjs';
 import { Logger, HttpService } from '../../core';
 import { Router } from '@angular/router';
-import { IAuthRequestSuccessResponse } from './auth.service';
+import { IAuthRequestSuccessResponse } from './authentication.service';
 import { WalletService } from './wallet.service';
 import { Store } from '@ngrx/store';
 import { AppState, selectUserState } from '../../app.states';
