+++ honestcash-v2/src/app/core/store/auth/auth.actions.ts
@@ -3,7 +3,7 @@ import {
   IAuthRequest,
   IAuthRequestSuccessResponse,
   IAuthRequestFailedResponse,
-} from '../../services/auth.service';
+} from '../../services/authentication.service';
 
 export enum AuthActionTypes {
   LOGIN = '[Auth] Login',
