+++ honestcash-v2/src/app/core/store/auth/auth.state.ts
@@ -1,8 +1,10 @@
 export interface State {
   // is a user authenticated?
   isAuthenticated: boolean;
+  token: string | null;
 }
 
 export const initialState: State = {
   isAuthenticated: false,
+  token: null,
 };
